rename a few var & files
